Migrate to libuild, flatten monorepo, add disposal protocol (3.1)
Build/structure:
- Replace Rollup + Jest with @b9g/libuild; tests via @b9g/libuild/test
- Flatten packages/* into single package with subpath entry points:
  repeater/core (class only), repeater/static (race/merge/zip/latest as
  named fns), repeater/timers, repeater/limiters. Default `repeater`
  re-exports the full class with statics attached (no breaking change).
- Deprecate pubsub (removed) and react-hooks (already gone)
- Switch to npm lockfile; TypeScript 5.x with isolatedModules

API:
- Add Symbol.asyncDispose / Symbol.dispose (ES2025) with runtime guards
- Rename aux exports: delay->createDelay, interval->createInterval,
  timeout->createTimeout, semaphore->createSemaphore, throttler->createThrottle
- createStop uses Object.assign; push() float-detection uses Object.create(nextP)

Fixes:
- index.ts re-exported type-only names as runtime bindings, crashing ESM
  imports; split into export/export type. isolatedModules now guards this.

Tests: 163/163 green on both node and bun.

Refs #81
