+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules
+dist
 lib
 coverage
 npm-debug.log*
