--- packages/limiters/package.json
@@ -1,56 +1,0 @@
-{
-  "name": "@repeaterjs/limiters",
-  "version": "0.3.3",
-  "description": "Basic async iterators functions for limiting concurrency",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/repeaterjs/repeater",
-    "directory": "packages/limiters"
-  },
-  "license": "MIT",
-  "files": [
-    "/lib"
-  ],
-  "main": "lib/limiters.cjs.js",
-  "module": "lib/limiters.esm.js",
-  "types": "lib/limiters.d.ts",
-  "exports": {
-    ".": {
-      "types": "./lib/limiters.d.ts",
-      "require": "./lib/limiters.cjs.js",
-      "default": "./lib/limiters.esm.js"
-    }
-  },
-  "scripts": {
-    "prebuild": "yarn run clean",
-    "build": "rollup -c ./rollup.config.js",
-    "clean": "shx rm -rf ./lib",
-    "lint": "eslint --ext ts src",
-    "prepublishOnly": "yarn run test && yarn run build",
-    "test": "jest --config ./jest.config.js --rootDir ./ --color"
-  },
-  "dependencies": {
-    "@repeaterjs/repeater": "^3.0.0",
-    "@repeaterjs/timers": "^0.3.2"
-  },
-  "devDependencies": {
-    "@types/jest": "^26.0.7",
-    "@typescript-eslint/eslint-plugin": "^3.7.1",
-    "@typescript-eslint/parser": "^3.7.1",
-    "eslint": "^7.5.0",
-    "eslint-config-prettier": "^6.4.0",
-    "eslint-config-recommended-plus-types": "^1.0.0",
-    "eslint-plugin-jest": "^23.20.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "jest": "^26.2.1",
-    "prettier": "^2.0.5",
-    "rollup": "^2.23.0",
-    "rollup-plugin-typescript2": "^0.27.1",
-    "shx": "^0.3.2",
-    "ts-jest": "^26.1.4",
-    "typescript": "^3.9.7"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
