--- packages/repeater/.eslintrc.json
@@ -1,43 +1,0 @@
-{
-  "env": {
-    "browser": true,
-    "commonjs": true,
-    "es6": true,
-    "node": true
-  },
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "sourceType": "module",
-    "project": "./tsconfig.json",
-    "createDefaultProgram": true
-  },
-  "extends": [
-    "eslint:recommended",
-    "recommended-plus-types",
-    "plugin:prettier/recommended",
-    "prettier/@typescript-eslint",
-    "plugin:jest/recommended"
-  ],
-  "plugins": [
-    "@typescript-eslint",
-    "jest",
-    "prettier"
-  ],
-  "rules": {
-    "prettier/prettier": [
-      "error",
-      {
-        "trailingComma": "all",
-        "arrowParens": "always"
-      }
-    ],
-    "linebreak-style": ["error", "unix"],
-    "@typescript-eslint/no-unused-vars": [
-      "error",
-      {
-        "argsIgnorePattern": "^_",
-        "varsIgnorePattern": "^_"
-      }
-    ]
-  }
-}
