--- packages/repeater/.gitignore
@@ -1,4 +1,0 @@
-cjs
-repeater.js
-*.d.ts
-*.js.map
