--- packages/repeater/package.json
@@ -1,61 +1,0 @@
-{
-  "name": "@repeaterjs/repeater",
-  "version": "3.0.6",
-  "description": "The missing constructor function for creating safe async iterators",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/repeaterjs/repeater",
-    "directory": "packages/repeater"
-  },
-  "license": "MIT",
-  "files": [
-    "repeater.js",
-    "*.d.ts",
-    "*.js.map",
-    "cjs"
-  ],
-  "type": "module",
-  "main": "cjs/repeater.js",
-  "module": "repeater.js",
-  "types": "repeater.d.ts",
-  "exports": {
-    ".": {
-      "require": {
-        "types": "./cjs/repeater.d.ts",
-        "default": "./cjs/repeater.js"
-      },
-      "types": "./repeater.d.ts",
-      "default": "./repeater.js"
-    }
-  },
-  "scripts": {
-    "prebuild": "yarn run clean",
-    "build": "rollup -c ./rollup.config.js",
-    "clean": "shx rm -rf cjs dist repeater.js *.d.ts *.js.map",
-    "lint": "eslint --ext ts src",
-    "prepublishOnly": "yarn run test && yarn run build",
-    "test": "jest --config ./jest.config.cjs --rootDir ./ --color"
-  },
-  "devDependencies": {
-    "@types/jest": "^26.0.7",
-    "@typescript-eslint/eslint-plugin": "^3.7.1",
-    "@typescript-eslint/parser": "^3.7.1",
-    "eslint": "^7.5.0",
-    "eslint-config-prettier": "^6.4.0",
-    "eslint-config-recommended-plus-types": "^1.0.0",
-    "eslint-plugin-jest": "^23.20.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "jest": "^26.2.1",
-    "magic-string": "^0.25.7",
-    "prettier": "^2.0.5",
-    "rollup": "^2.23.0",
-    "rollup-plugin-typescript2": "^0.27.1",
-    "shx": "^0.3.2",
-    "ts-jest": "^26.1.4",
-    "ts-transform-import-path-rewrite": "^0.2.1",
-    "typescript": "^3.9.7"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
