--- packages/repeater/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "./tsconfig.base.json",
-  "files": [
-    "src/repeater.ts"
-  ]
-}
