--- packages/timers/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@repeaterjs/timers",
-  "version": "0.3.3",
-  "description": "Async iterator timer functions",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/repeaterjs/repeater",
-    "directory": "packages/timers"
-  },
-  "license": "MIT",
-  "files": [
-    "/lib"
-  ],
-  "main": "lib/timers.cjs.js",
-  "module": "lib/timers.esm.js",
-  "types": "lib/timers.d.ts",
-  "exports": {
-    ".": {
-      "types": "./lib/timers.d.ts",
-      "require": "./lib/timers.cjs.js",
-      "default": "./lib/timers.esm.js"
-    }
-  },
-  "scripts": {
-    "prebuild": "yarn run clean",
-    "build": "rollup -c ./rollup.config.js",
-    "clean": "shx rm -rf ./lib",
-    "lint": "eslint --ext ts src",
-    "prepublishOnly": "yarn run test && yarn run build",
-    "test": "jest --config ./jest.config.js --rootDir ./ --color"
-  },
-  "dependencies": {
-    "@repeaterjs/repeater": "^3.0.0"
-  },
-  "devDependencies": {
-    "@types/jest": "^26.0.7",
-    "@typescript-eslint/eslint-plugin": "^3.7.1",
-    "@typescript-eslint/parser": "^3.7.1",
-    "eslint": "^7.5.0",
-    "eslint-config-prettier": "^6.4.0",
-    "eslint-config-recommended-plus-types": "^1.0.0",
-    "eslint-plugin-jest": "^23.20.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "jest": "^26.2.1",
-    "prettier": "^2.0.5",
-    "rollup": "^2.23.0",
-    "rollup-plugin-typescript2": "^0.27.1",
-    "shx": "^0.3.2",
-    "ts-jest": "^26.1.4",
-    "typescript": "^3.9.7"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
