--- packages/timers/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "./tsconfig.base.json",
-  "files": [
-    "src/timers.ts"
-  ]
-}
