+++ src/index.ts
@@ -1,0 +1,25 @@
+export {
+  RepeaterOverflowError,
+  FixedBuffer,
+  SlidingBuffer,
+  DroppingBuffer,
+  MAX_QUEUE_LENGTH,
+} from "./core.js";
+export type {
+  RepeaterBuffer,
+  Push,
+  Stop,
+  RepeaterExecutor,
+} from "./core.js";
+
+export { race, merge, zip, latest } from "./static.js";
+
+import { Repeater as RepeaterCore } from "./core.js";
+import { race, merge, zip, latest } from "./static.js";
+
+export class Repeater<T, TReturn = any, TNext = unknown> extends RepeaterCore<T, TReturn, TNext> {
+  static race = race;
+  static merge = merge;
+  static zip = zip;
+  static latest = latest;
+}
