+++ src/test/_testutils.ts
@@ -1,4 +1,5 @@
-import { Repeater, FixedBuffer } from "../repeater";
+import { Repeater, FixedBuffer } from "../index.js";
+
 // take from https://2ality.com/2019/07/testing-static-types.html
 export type AssertTypeEquals<T, TExpected> = T extends TExpected
   ? TExpected extends T
@@ -40,6 +41,21 @@ export async function* hangingGen<T = never>(): AsyncIterableIterator<T> {
   yield (Infinity as unknown) as T;
 }
 
+// Awaits a promise expected to reject and returns the rejection reason. Used
+// instead of expect(p).rejects in timing-sensitive tests, because the bun test
+// runner's .resolves/.rejects matchers insert different microtask ticks than a
+// plain await, which shifts when a repeater's floating-rejection propagation
+// lands relative to subsequent next() calls.
+export async function rejection(p: Promise<unknown>): Promise<unknown> {
+  try {
+    await p;
+  } catch (err) {
+    return err;
+  }
+
+  throw new Error("Expected promise to reject");
+}
+
 export function delayPromise<T>(
   wait: number,
   value?: T,
@@ -48,7 +64,7 @@ export function delayPromise<T>(
   return new Promise<T>((resolve, reject) => {
     setTimeout(() => {
       if (error == null) {
-        resolve(value);
+        resolve(value!);
       } else {
         reject(error);
       }
