+++ src/test/limiters.ts
@@ -1,8 +1,10 @@
-import { semaphore, throttler } from "../limiters";
+import { describe, test, expect } from "@b9g/libuild/test";
+import * as Sinon from "sinon";
+import { createSemaphore, createThrottle } from "../limiters.js";
 
 describe("limiters", () => {
-  test("semaphore", async () => {
-    const tokens = semaphore(4);
+  test("createSemaphore", async () => {
+    const tokens = createSemaphore(4);
     let i = 0;
     for await (const token of tokens) {
       expect(token.id).toBe(i);
@@ -15,8 +17,8 @@ describe("limiters", () => {
     }
   });
 
-  test("semaphore remaining", async () => {
-    const tokens = semaphore(4);
+  test("createSemaphore remaining", async () => {
+    const tokens = createSemaphore(4);
     const t1 = (await tokens.next()).value;
     expect(t1.remaining).toEqual(3);
     const t2 = (await tokens.next()).value;
@@ -32,13 +34,14 @@ describe("limiters", () => {
     t3.release();
     const t6 = (await tokens.next()).value;
     expect(t6.remaining).toEqual(1);
+    await tokens.return();
   });
 
-  test("throttler", async () => {
+  test("createThrottle", async () => {
     let prev = Date.now();
     let i = 0;
     const wait = 200;
-    for await (const _ of throttler(wait, { limit: 8 })) {
+    for await (const _ of createThrottle(wait, { limit: 8 })) {
       const next = Date.now();
       if (i !== 0 && i % 8 === 0) {
         expect(prev + wait).toBeCloseTo(next, -1.5);
@@ -53,35 +56,38 @@ describe("limiters", () => {
     }
   });
 
-  test("throttler timer window slides", async () => {
+  test("createThrottle timer window slides", async () => {
     let prev = Date.now();
     let i = 0;
     const wait = 200;
-    const spy = jest.spyOn(window, "setTimeout");
-    for await (const _ of throttler(wait, { limit: 8 })) {
-      const next = Date.now();
-      expect(prev).toBeCloseTo(next, -1.5);
-      if (i % 8 === 0) {
-        await new Promise((resolve) => setTimeout(resolve, wait * 2));
+    const spy = Sinon.spy(globalThis, "setTimeout");
+    try {
+      for await (const _ of createThrottle(wait, { limit: 8 })) {
+        const next = Date.now();
+        expect(prev).toBeCloseTo(next, -1.5);
+        if (i % 8 === 0) {
+          await new Promise((resolve) => setTimeout(resolve, wait * 2));
+        }
+
+        if (i >= 40) {
+          break;
+        }
+
+        i++;
+        prev = Date.now();
       }
 
-      if (i >= 40) {
-        break;
-      }
-
-      i++;
-      prev = Date.now();
+      // called 6 times by delay timer, 6 times by the promise
+      expect(spy.callCount).toBe(12);
+    } finally {
+      spy.restore();
     }
-
-    // called 6 times by delay timer
-    // called 6 times by the promise
-    expect(spy).toHaveBeenCalledTimes(12);
   });
 
-  test("throttler token reset", async () => {
+  test("createThrottle token reset", async () => {
     let i = 0;
     const wait = 200;
-    for await (const token of throttler(wait, { limit: 8 })) {
+    for await (const token of createThrottle(wait, { limit: 8 })) {
       expect(token.reset).toBeCloseTo(Date.now() + wait, -1.5);
       if (i >= 40) {
         break;
@@ -91,11 +97,11 @@ describe("limiters", () => {
     }
   });
 
-  test("throttler token remaining", async () => {
+  test("createThrottle token remaining", async () => {
     let remaining = 8;
     let i = 0;
     const wait = 200;
-    for await (const token of throttler(wait, { limit: 8 })) {
+    for await (const token of createThrottle(wait, { limit: 8 })) {
       remaining--;
       expect(token.remaining).toEqual(remaining);
       if (token.remaining === 0) {
@@ -109,11 +115,11 @@ describe("limiters", () => {
     }
   });
 
-  test("throttler with cooldown", async () => {
+  test("createThrottle with cooldown", async () => {
     let i = 0;
     const wait = 200;
     let prev = Date.now();
-    for await (const _ of throttler(wait, { cooldown: true })) {
+    for await (const _ of createThrottle(wait, { cooldown: true })) {
       const next = Date.now();
       expect(prev + wait).toBeCloseTo(next, -1.5);
       if (i >= 4) {
@@ -125,17 +131,19 @@ describe("limiters", () => {
     }
   });
 
-  test("throttler cleans up", async () => {
+  test("createThrottle cleans up", async () => {
+    const setTimeoutSpy = Sinon.spy(globalThis, "setTimeout");
+    const clearTimeoutSpy = Sinon.spy(globalThis, "clearTimeout");
     try {
-      jest.useFakeTimers();
-      const throttle = throttler(200, { limit: 8 });
+      const throttle = createThrottle(200, { limit: 8 });
       await throttle.next();
-      expect(setTimeout).toHaveBeenCalledTimes(1);
+      expect(setTimeoutSpy.callCount).toBe(1);
       await expect(throttle.return()).resolves.toEqual({ done: true });
-      expect(clearTimeout).toHaveBeenCalledTimes(1);
+      expect(clearTimeoutSpy.callCount).toBe(1);
       await expect(throttle.next()).resolves.toEqual({ done: true });
     } finally {
-      jest.useRealTimers();
+      setTimeoutSpy.restore();
+      clearTimeoutSpy.restore();
     }
   });
 });
